requirements.txt
fs
fs-s3fs
Django
djangorestframework
psycopg2
drf-spectacular